requirements.txt
selenium
requests
pytest
allure-pytest
pytest-xdist
flake8
